+++ lib/components/DateTime/DateField.tsx
@@ -63,7 +63,7 @@ export interface DateFieldProps extends React.Props<DateFieldType> {
     onExpand?: (expanded: boolean) => void;
 
     /** React node to render at the far side of the label. */
-    labelExtraAction?: React.ReactNode;
+    labelFarSide?: React.ReactNode;
 
     attr?: DatePickerAttributes & FormFieldAttributes;
 }
@@ -115,7 +115,7 @@ export const DateField: React.StatelessComponent<DateFieldProps> = (props: DateF
             className={props.className}
             attr={fieldAttr}
             tooltip={props.tooltip}
-            labelExtraAction={props.labelExtraAction}
+            labelFarSide={props.labelFarSide}
         >
             <DatePicker
                 name={props.name}