+++ lib/components/Field/ComboField.tsx
@@ -111,7 +111,7 @@ export interface ComboFieldProps extends React.Props<ComboFieldType> {
     /** Classname to append to top level element of ComboInput */
     inputClassName?: string;
     /** React node to render at the far side of the label. */
-    labelExtraAction?: React.ReactNode;
+    labelFarSide?: React.ReactNode;
 
     attr?: ComboInputAttributes & FormFieldAttributes;
 }
@@ -183,7 +183,7 @@ export const ComboField: React.StatelessComponent<ComboFieldProps> = (props: Com
             tooltip={props.tooltip}
             className={props.className}
             attr={fieldAttr}
-            labelExtraAction={props.labelExtraAction}
+            labelFarSide={props.labelFarSide}
         >
             <div>
                 <ComboInput