+++ lib/components/Field/RadioField.tsx
@@ -53,7 +53,7 @@ export interface RadioFieldProps extends React.Props<RadioFieldType> {
     /** Classname to append to top level element of RadioInput */
     inputClassName?: string;
     /** React node to render at the far side of the label. */
-    labelExtraAction?: React.ReactNode;
+    labelFarSide?: React.ReactNode;
 
     attr?: RadioInputAttributes & FormFieldAttributes;
 }
@@ -141,7 +141,7 @@ export const RadioField: React.StatelessComponent<RadioFieldProps> = (props: Rad
             tooltip={props.tooltip}
             className={props.className}
             attr={fieldAttr}
-            labelExtraAction={props.labelExtraAction}
+            labelFarSide={props.labelFarSide}
         >
             <div>
                 {options}