+++ lib/components/Field/ToggleField.tsx
@@ -44,7 +44,7 @@ export interface ToggleFieldProps extends React.Props<ToggleFieldType> {
     /** Classname to append to top level element of SelectInput */
     inputClassName?: string;
     /** React node to render at the far side of the label. */
-    labelExtraAction?: React.ReactNode;
+    labelFarSide?: React.ReactNode;
 
     attr?: FormFieldAttributes & ToggleAttributes;
 }
@@ -90,7 +90,7 @@ export const ToggleField: React.StatelessComponent<ToggleFieldProps> = (props: T
             tooltip={props.tooltip}
             className={props.className}
             attr={fieldAttr}
-            labelExtraAction={props.labelExtraAction}
+            labelFarSide={props.labelFarSide}
         >
             <Toggle
                 on={props.value}